codebrew/requirements.txt
python-json-logger>=2.0.7
rich>=13.7.0
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
pydantic>=2.5.2
python-dotenv>=1.0.0
cachetools>=5.3.2
groq>=0.4.1
openai>=1.3.7
cohere>=4.37
google-generativeai>=0.3.1
together>=0.2.8
litellm>=1.30.3
Pillow>=10.1.0
requests>=2.31.0
aiohttp>=3.9.1
asyncio>=3.4.3
httpx>=0.25.2
tenacity>=8.2.3
black>=23.11.0
isort>=5.12.0
mypy>=1.7.1
pytest>=7.4.3
pytest-asyncio>=0.23.2
pytest-cov>=4.1.0